+++ lib/vagrant_mutagen_utilizer/action/remove_config.rb
@@ -1,13 +1,13 @@
 # frozen_string_literal: true
 
-require_relative '../mutagen'
+require_relative '../orchestrator'
 
 module Vagrant
   module Mutagen
     module Utilize
       module Action
-        # Start mutagen project
-        class StartOrchestration
+        # Remove SSH config entry from user ssh config file
+        class RemoveConfig
           def initialize(app, env)
             @app = app
             @machine = env[:machine]
@@ -18,8 +18,8 @@ module Vagrant
           def call(env)
             return unless @config.orchestrate?
 
-            m = Mutagen.new(@machine, @console)
-            m.start_orchestration
+            o = Orchestrator.new(@machine, @console)
+            o.remove_ssh_config_entry
             @app.call(env)
           end
         end
